requirements.txt
streamlit==1.40.2
streamlit-chat==0.1.1
ollama==0.1.8
pyyaml==6.0.2
streamlit-authenticator==0.4.1